+++ Changelog.md
@@ -21,6 +21,7 @@ Breaking Changes:
  * Type System: Explicit conversions between two types are disallowed if it changes more than one of sign, width or kind at the same time.
  * Type System: Explicit conversions from literals to enums are only allowed if the value fits in the enum.
  * Type System: Explicit conversions from literals to integer type is as strict as implicit conversions.
+ * Type System: Support ``address(...).codehash`` to retrieve the codehash of an account.
  * Type System: Unary negation can only be used on signed integers, not on unsigned integers.
 
 Language Features:
@@ -156,7 +157,6 @@ Bugfixes:
 
 Important Bugfixes:
  * Type Checker: Disallow two or more free functions with identical name (potentially imported and aliased) and parameter types.
-
 
 Compiler Features:
  * Export compiler-generated utility sources via standard-json or combined-json.